+++ tools/testnet/src/main.rs
@@ -2,8 +2,6 @@ use clap::Parser;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    libra_testnet::testnet_cli::TestnetCli::parse()
-        .run()
-        .await?;
+    libra_testnet::cli_main::TestnetCli::parse().run().await?;
     Ok(())
 }
